+++ tests/test_template.cpp
@@ -1,6 +1,5 @@
 //fusa:test REQ-TMPL001
 //fusa:test REQ-TMPL002
-//fusa:test REQ-TMPL003
 //fusa:test REQ-TMPL004
 #include <catch2/catch_all.hpp>
 #include "template/template_gen.hpp"